added plug.vim
plug.vim is a preferred plugin manager to vim-pathogen. plug.vim works on Vim, so it has been included into the '.vim/vimrc'

in the case that plug.vim doesn't work, it should be easy to switch to vim-pathogen by deleting the plug.vim file and cloning all plugins into the '.vim/bundle' directory.